+++ Examples/Write-ProgressEx.parm.ps1
@@ -1,14 +1,17 @@
-# mazzy@mazzy.ru, 2017-10-10
+# mazzy@mazzy.ru, 2017-10-14
 # https://github.com/mazzy-ax/write-progressEx
 
-Import-Module -Force "..\Write-ProgressEx.psd1"
+$me = Split-Path -Leaf $PSCommandPath
+$path = Split-Path -Parent $PSCommandPath
+$module = Join-Path (Split-Path -Parent $path) "Write-ProgressEx.psd1"
+Import-Module -Force $module
 
 $nodes = 1..20
 $names = 1..50
 
-write-ProgressEx "parm nodes" -Total $nodes -ShowMessages
+write-ProgressEx "parm nodes" -Total $nodes.Count -ShowMessagesOnEnd
 $nodes | Where-Object {$true} | ForEach-Object {
-    write-ProgressEx "parm names" -Total $names -id 1
+    write-ProgressEx "parm names" -Total $names.Count -id 1
     $names | Where-Object {$true} | ForEach-Object {
         #...
         write-ProgressEx -id 1 -increment -Status "inner"